+++ vite.config.js
@@ -1,14 +1,22 @@
 /** @type {import('vite').UserConfig} */
 
 import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
+import { tanstackRouter } from '@tanstack/router-plugin/vite';
 import tailwindcss from '@tailwindcss/vite';
+import react from '@vitejs/plugin-react';
 
 export default defineConfig(() => {
   return {
     build: {
       outDir: 'build',
     },
-    plugins: [react(), tailwindcss()],
+    plugins: [
+      tanstackRouter({
+        target: 'react',
+        autoCodeSplitting: true,
+      }),
+      react(),
+      tailwindcss(),
+    ],
   };
 });